Allow symbols for credential flow
Because of my production environment, adding a JSON file a little bit of
a pain. However: using the intializer it's readily possible to set my
credentials manually from my environment. I noticed that while all the
credential details can be either a String or a Symbol, the 'flow' (i.e.,
`web`) **must** be a string.

This allows it to accept both a string and a symbol.