Fixes broken filters URL reference
Because of the line wrap in the Rubydoc, the URL for filters documentation ends up with an extra %20A in it, which breaks the link from Rubydoc. This undoes the line break so the documentation link is correct. I left it pointing to the V3 docs, since the V4 docs don't seem to have a reference section yet.