Improve language in Logging section of README (#642)
For readers following the README's sequence of instructions to get logging to STDOUT/STDERR, they will generally add code in the order which it appears in the README. That results in this:
```
Google::Apis.logger.level = Logger::DEBUG
Google::Apis.logger = Logger.new(STDERR)
```

However, this configuration will not have the desired effect in Rails apps, since the newly created logger will not respect the log level which was actually set on the Rails logger. By changing the order in which these code snippets appear in the README, we can help readers avoid this gotcha.

Additionally, I offer a small semantic improvement to help readers better understand the two options they have for changing the default logger.