storage: Optimize memory usage when upload chunk size is set to 0
https://github.com/googleapis/google-api-ruby-client/pull/13283 added
support for disabling upload chunking. However, when chunking is
disabled, the upload attempts to read the entire file in memory
instead of passing along the IO object. This often leads to
out-of-memory errors when the uploaded file is large. We can
significantly reduce memory by passing along the unfettered upload IO
object.

Closes #13340